# adodasopenapi.yaml
openapi: 3.0.3
info:
  title: Adodas API
  description: |
    API REST completa para gestión de **clientes, ítems, pedidos, reembolsos y servicios**.
    
    ## ✨ Características:
    - ✅ **CRUD completo** todos los recursos
    - ✅ **Validaciones robustas** (DNI, emails, precios)
    - ✅ **Filtros avanzados** por estado, precio, cliente
    - ✅ **Errores específicos** 400/404/500 diferenciados
    - ✅ **Ejemplos reales** en todas las respuestas
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Desarrollo Local
...
components:
  schemas:
    ErrorResponseDto:
      type: object
      properties:
        code:
          type: integer
          description: Código HTTP del error
        message:
          type: string
          description: Descripción del error
      required: [ code, message ]
...
    ItemDto:
      type: object
      properties:
        name:
          type: string
          example: "Camiseta Nike Sport"
        brand:
          type: string
          example: "Nike"
        price:
          type: number
          format: float
          minimum: 0
          example: 29.99
        category:
          $ref: '#/components/schemas/CategoryEnum'
        stock:
          type: integer
          minimum: 0
          example: 150
        description:
          type: string
          example: "Camiseta de algodón 100% transpirable"
        createdDate:
          type: string
          format: date-time
          readOnly: true
          example: "2026-01-15T10:30:00Z"
    CategoryEnum:
      type: string
      enum: [ SHOES, TROUSERS, SWEATER, SHIRTS ]
      example: "SHIRTS"
...
paths:
...
  /items:
    get:
      tags: [ ITEMS ]
      summary: Lista todos los ítems
      description: Obtiene una lista de ítems con filtros opcionales por marca, categoría o precio.
      parameters:
        - name: brand
          in: query
          required: false
          schema:
            type: string
          description: Filtra por marca del ítem.
        - name: category
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/CategoryEnum'
          description: Filtra por categoría del ítem.
        - name: price
          in: query
          required: false
          schema:
            type: number
            format: float
          description: Filtra por precio exacto del ítem.
      responses:
        '200':
          description: Lista de ítems obtenida correctamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ItemDto'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    post:
      tags: [ ITEMS ]
      summary: Crea un nuevo ítem
      description: Registra un nuevo ítem en el catálogo validando precio y categoría.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemDto'
      responses:
        '201':
          description: Ítem creado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDto'
        '400':
          description: Error de validación en los datos proporcionados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 400
                message: "Validation failed"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"
  /items/{itemId}:
    get:
      tags: [ ITEMS ]
      summary: Obtiene un ítem por su ID
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del ítem.
      responses:
        '200':
          description: Ítem encontrado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDto'
        '404':
          description: Ítem no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Ítem no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    put:
      tags: [ ITEMS ]
      summary: Modifica un ítem existente
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del ítem a modificar.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ItemDto'
      responses:
        '200':
          description: Ítem modificado correctamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDto'
        '400':
          description: Error de validación en los datos proporcionados.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 400
                message: "Validation failed"
        '404':
          description: Ítem no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Ítem no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    patch:
      tags: [ ITEMS ]
      summary: Actualiza parcialmente un ítem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del ítem a actualizar parcialmente.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              example:
                price: 35.99
                stock: 200
      responses:
        '200':
          description: Ítem actualizado parcialmente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ItemDto'
        '400':
          description: Error de validación o cuerpo inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 400
                message: "Validation failed"
        '404':
          description: Ítem no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Ítem no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"

    delete:
      tags: [ ITEMS ]
      summary: Elimina un ítem
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: ID del ítem a eliminar.
      responses:
        '204':
          description: Ítem eliminado correctamente (sin contenido en la respuesta).
        '404':
          description: Ítem no encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 404
                message: "Ítem no encontrado"
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
              example:
                code: 500
                message: "Error interno del servidor"
...